[#5] Feat: 중복 ID 조회 API 구현
- /users/duplicate/id/{userId} 로 요청을 보내는 경우, URL에 전달된 id가 존재하는지 여부를 boolean으로 리턴하도록 구현

- select * 로 조회하는 것이 안좋다고 해서 user_id 컬럼을 작성했는데 왜 *가 안좋은지 조사해봐야 할듯.
