--- BE/src/main/java/com/codesquad/team1/signup/domain/UserRepository.java
@@ -1,6 +1,0 @@
-package com.codesquad.team1.signup.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Integer> {
-}
